Extract vehicle date utilities
